refactor: move PasswordInput to the components directory
